backend/pyproject.toml
[project]
name = "cargoletsalerts"
version = "0.1.0"
description = "Alert dashboard for medical product supplier analytics"
requires-python = ">=3.11"
dependencies = [
    "django>=5.0,<6.0",
    "djangorestframework>=3.15,<3.17",
    "django-cors-headers>=4.3",
    "django-filter>=24.1",
    "python-dotenv>=1.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-django>=4.8",
    "ruff>=0.4",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.ruff]
line-length = 88
target-version = "py311"
